move functionals to utils normalize
